+++ Assets/TK2DROOT/tk2d/Shaders/CutoutVertexColor.shader
@@ -36,7 +36,7 @@ Shader "tk2d/CutoutVertexColor"
 			{
 				float4 vertex : POSITION;
 				fixed4 color : COLOR;
-				half2 texcoord : TEXCOORD0;
+				float2 texcoord : TEXCOORD0;
 			};
 
 			v2f_vct vert_vct(vin_vct v)
