+++ Assets/TK2DROOT/tk2d/Shaders/LitSolidVertexColor.shader
@@ -14,7 +14,7 @@ Shader "tk2d/LitSolidVertexColor"
 		CGPROGRAM
 		#pragma surface surf Lambert
 		struct Input {
-			half2 uv_MainTex;
+			float2 uv_MainTex;
 			fixed4 color : COLOR;
 		};
 		sampler2D _MainTex;
